Added an "onSubmit" notification / verification option.
This allows the plugin, to post a message as soon as jobs are submitted.
Previously, it was only able to send a message, when the jobs were actually
started.

If there is a long queue, this meant that verification messages were highly
delayed.

Do note that this message is only reported once per MemoryState and only
reset if you issue a retrigger.